add route and trip args to selectData, add selectTrips function
These help toward #4 so you can now get a list of trips for a given
device+date then select on them. The view is still fairly slow, even
with indexes, so may want to create rlev as a table. rlev as a view
does not get returned by getTables so the call to that function in
getColumns has been commented.