plots should take recs, not curs
replace plotDeviceDate with plotCoords, removing call to
io.selectData. The plot function should just receive a list of points
(or for now a rec array) and not a cursor to a db object. When trying
to make multiple plots with the same data, you don't want to have to
query the db each time.