split io.py into copyRemoteDB and eventsDBManager
copyRemoteDB contains the functions used to download the remote
PostgreSQL db and construct the local SQLite copy. This is only saved
to show how the SQLite db was created so can be kept separate from the
other code in development.

eventsDBManager is based on Jacob's gtfsDbManager and contains the
EventsDB class which has several methods to generate and execute
queries to the SQLite db. Rather than having a cursor and connection
object running around, those are now attributes of the class.

Cleaned up some of the specialized selection methods to call the more
general selectData and added distinct arg to the latter. Split off
code that converts cursor rows to numpy recarray into rowsToRec
function.