add metersBetweenLatLonPair and printCurrentTime functions
metersBetweenLatLonPair is a copy of Jacob Lynn's version in pygtfs,
but returns meters instead of km and uses numpy instead of math
functions. It implements a great circle distance approximation and is
vectorized so it works much faster than geopy on large lists.

printCurrentTime just prints the ISO formatted time for quick
debugging/logging.