speed up time and distance conversions, clean encoding, validate knn
Add several functions that should duplicate calculations
(unixmsToWeekSecs, lanlonToMeters, etc) but perform much faster with
vectorization. These need to be tested for consistency and speed, and
the slower ones can be deleted. In particular, the new time functions
have different (probably incorrect) timezones and the start of the
week is Thursday.

Split off the label encoding from preprocess into its own
function. KNN doesn't like it when the test set has labels (trip_ids)
that aren't found in the training set, so now we join the test and
training labels in the encoder object.

Once KNN is fit, use its built-in score method to check accuracy and
use the metrics submodule to generate and plot the confusion matrix.