convert timezone in df returned by selectData to mainland Spain
San Sebastian is in Europe/Madrid timezone (aka CEST), usually 1hr
ahead of UTC (but it also goes on daylight savings time in
summer). Unix ms is stored in UTC timezone so we first identify values
as UTC then convert to CEST for display which is useful for comparing
against the HHMM departure time in the parsed trip IDs.

Note - SQL query is still done in UTC unix time, so care must be taken
when the difference between these two matters.