simply Sequence class, flesh out getSequences, remove dtMin
replace Sequence method addPoint by just passing full arrays to
__init__. getSequences now slices the data to the appropriate length
and fills the sequence object arrays, and generates an array of
sequences (and then a list of those arrays!). This organization is a
bit confusion and may need some cleanup, but the data are now chunked
as desired.

Also removed dtMin option everywhere in favor of nPts for simplicity.