panda-fy preprocess, change output of getSequences from dict to dataframe
preprocess now returns both the design matrix (xData) and the
corresponding labels (which still have to be encoded by
encodeLabels). Both are now pandas dataframes instead of ndarrays.

Add nPts arg to preprocess and classify. In preprocess, this
determines whether to use scaled times and distances as the feature
vector or to use time series of length nPts (if nPts > 1).

Still need to feed nPts>1 option into classify and tell kNN the
relevant distance measure for comparing sequences or to call DTW
library.