add vectorized DTW classifier, fix in getSequences to cull short trips
Open-ended DTW requires query sequence to be shorter than templace
sequence, so we have to remove trips that are shorter than nPts long.

Started vecDTWClassifier to call dtw.vecRDTW - still need to test that
function to see if results are ok, but for now let's see how well the
classifier does.