remove getAccuracy
assignedTripDict is not a cumulative list of assigned trips, so is not
a good data structure for measuring trip assignment accuracy. This
problem requires one to know what the actual trip_id is and compare it
to the assigned value, for each location/event/sequence of the
former. assignedTripDict calls its pop method in several places where
bad trips get dropped from the dict, and Null trips aren't assigned to
the dict.