Fix projection, and smoothly handle transitions between trips.
1. Projection now minimizes perpendicular distance to route, subject to
condition that the deviation from the target postmile is less than 200m.
(Previously we were minimizing perpDist + postDev, which has the property of
pushing the bus forward towards the next segment in the shape file.)

2. When transitioning between trips, if we haven't yet arrived at the first
stop on a trip, include the shapepoints of the previous trip in the block as
part of the projection. Allow negative postmile and target postmile in this
case.